backend/requirements.txt
fastapi==0.95.2 
uvicorn==0.22.0 
pydantic==1.10.13 
starlette==0.27.0 
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
Pillow==10.1.0
numpy==1.25.2